legal_assistant_chatbot/requirements.txt
fpdf
groq
gradio
pymupdf
pillow
pytesseract